requirements.txt
matrix-nio[e2e]==0.25.2
aiohttp==3.10.11
aiosqlite==0.20.0
psutil==5.9.8
PyYAML==6.0.2
